+++ SW_API_Kotlin_Project/app/src/main/java/com/sw/sw_api_kotlin_project/screen/base/BaseViewModel.kt
@@ -1,7 +1,7 @@
 package com.sw.sw_api_kotlin_project.screen.base
 
 import androidx.lifecycle.ViewModel
-import com.sw.sw_api_kotlin_project.model.entity.PageType
+import com.sw.sw_api_kotlin_project.data.model.entity.PageType
 
 abstract class BaseViewModel : ViewModel() {
     var page: Int = 1
