+++ SW_API_Kotlin_Project/app/src/main/java/com/sw/sw_api_kotlin_project/screen/home/HomeAdapter.kt
@@ -7,8 +7,8 @@ import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.sw.sw_api_kotlin_project.R
-import com.sw.sw_api_kotlin_project.model.entity.HomeItem
-import com.sw.sw_api_kotlin_project.model.entity.ListType
+import com.sw.sw_api_kotlin_project.data.model.entity.HomeItem
+import com.sw.sw_api_kotlin_project.data.model.entity.ListType
 
 class HomeAdapter(
     private val homeItemList: List<HomeItem>,
