+++ SW_API_Kotlin_Project/app/src/main/java/com/sw/sw_api_kotlin_project/screen/home/HomeFragment.kt
@@ -10,9 +10,9 @@ import androidx.navigation.fragment.findNavController
 import com.google.android.material.appbar.MaterialToolbar
 import com.sw.sw_api_kotlin_project.R
 import com.sw.sw_api_kotlin_project.screen.base.BaseFragment
-import com.sw.sw_api_kotlin_project.model.entity.HomeItem
+import com.sw.sw_api_kotlin_project.data.model.entity.HomeItem
 import com.sw.sw_api_kotlin_project.databinding.FragmentHomeBinding
-import com.sw.sw_api_kotlin_project.model.entity.ListType
+import com.sw.sw_api_kotlin_project.data.model.entity.ListType
 import dagger.hilt.android.AndroidEntryPoint
 
 /**
