+++ SW_API_Kotlin_Project/app/src/main/java/com/sw/sw_api_kotlin_project/screen/others/OthersFragment.kt
@@ -10,7 +10,7 @@ import androidx.navigation.fragment.findNavController
 import com.google.android.material.appbar.MaterialToolbar
 import com.sw.sw_api_kotlin_project.R
 import com.sw.sw_api_kotlin_project.screen.base.BaseFragment
-import com.sw.sw_api_kotlin_project.model.entity.WebViewInfo
+import com.sw.sw_api_kotlin_project.data.model.entity.WebViewInfo
 import com.sw.sw_api_kotlin_project.databinding.FragmentOthersBinding
 import dagger.hilt.android.AndroidEntryPoint
