+++ SW_API_Kotlin_Project/data/src/main/java/com/sw/sw_api_kotlin_project/data/database/FavoriteDao.kt
@@ -1,10 +1,10 @@
-package com.sw.sw_api_kotlin_project.database
+package com.sw.sw_api_kotlin_project.data.database
 
 import androidx.room.Dao
 import androidx.room.Delete
 import androidx.room.Insert
 import androidx.room.Query
-import com.sw.sw_api_kotlin_project.model.entity.Favorite
+import com.sw.sw_api_kotlin_project.data.model.entity.Favorite
 import kotlinx.coroutines.flow.Flow
 
 @Dao
