+++ SW_API_Kotlin_Project/data/src/main/java/com/sw/sw_api_kotlin_project/data/database/FavoriteDatabase.kt
@@ -1,9 +1,9 @@
-package com.sw.sw_api_kotlin_project.database
+package com.sw.sw_api_kotlin_project.data.database
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
-import com.sw.sw_api_kotlin_project.model.entity.Favorite
+import com.sw.sw_api_kotlin_project.data.model.entity.Favorite
 
 @Database(entities = [Favorite::class], version = 1, exportSchema = false)
 @TypeConverters(Converters::class)
