+++ SW_API_Kotlin_Project/data/src/main/java/com/sw/sw_api_kotlin_project/data/model/entity/Favorite.kt
@@ -1,4 +1,4 @@
-package com.sw.sw_api_kotlin_project.model.entity
+package com.sw.sw_api_kotlin_project.data.model.entity
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
